Rename test.home app to test.example app
