+++ laces/test/settings.py
@@ -35,7 +35,7 @@ ALLOWED_HOSTS = ["localhost", "testserver"]
 INSTALLED_APPS = [
     "laces",
     "laces.test",
-    "laces.test.home",
+    "laces.test.example",
     "laces.test.components",
     "django.contrib.admin",
     "django.contrib.auth",
