+++ laces/test/urls.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.urls import path
 
-from laces.test.home.views import home
+from laces.test.example.views import home
 
 
 urlpatterns = [
